supernova_meteo/requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
backports.zoneinfo;python_version<"3.9"
billiard==3.6.4.0
bondpy
celery==5.2.7
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.1.3
django-appconf==1.0.5
django-celery==3.1.17
django-compressor==4.1
django-compressor-toolkit==0.6.0
idna==3.4
kombu==5.2.4
prompt-toolkit==3.0.36
pycrypto==2.6.1
pytz==2022.6
rcssmin==1.1.0
redis==4.4.0
requests==2.28.1
rjsmin==1.2.0
schedule==1.1.0
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5

